hook to bulk delete and functionality added

+++ components/data-table.tsx
@@ -23,6 +23,7 @@ import {
 import {Button} from "@/components/ui/button";
 import {Input} from "@/components/ui/input";
 import {Trash} from "lucide-react";
+import { useConfirm } from '@/hooks/use-confirm';
 
 interface DataTableProps<TData, TValue> {
     columns: ColumnDef<TData, TValue>[];
@@ -39,6 +40,11 @@ export function DataTable<TData, TValue>({
                                             onDelete,
                                             disabled,
                                         }: DataTableProps<TData, TValue>) {
+    const [ConfirmDialog, confirm] = useConfirm(
+        "Are you Sure",
+        "You are about to perform a bulk delete"
+    )
+
     const [sorting, setSorting] = React.useState<SortingState>([]);
     const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>([]);
     const [rowSelection, setRowSelection ] = React.useState({});
@@ -61,6 +67,7 @@ export function DataTable<TData, TValue>({
 
     return (
         <div>
+            <ConfirmDialog />
             <div className="flex items-center py-4">
                 <Input
                     placeholder={`Filter ${filterKey} ...`}
@@ -75,9 +82,13 @@ export function DataTable<TData, TValue>({
                             size="sm" 
                             variant="outline" 
                             className="ml-auto font-normal text-xs"
-                            onClick={() =>{
-                                onDelete(table.getFilteredSelectedRowModel().rows)
-                                table.resetRowSelection()
+                            onClick={async () =>{
+                                const ok = await confirm();
+                                if(ok) {
+                                    onDelete(table.getFilteredSelectedRowModel().rows)
+                                    table.resetRowSelection()
+                                }
+
                             }}
                     >
                         <Trash  className="size-4 mr-2"/> Delete ({table.getFilteredSelectedRowModel().rows.length})
